+++ projects/egeo-demo/src/app/modules/demos/demo.module.ts
@@ -14,7 +14,6 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { DemoLayoutComponent } from './demo-layout/demo-layout';
 import { SharedModule } from '../../shared/shared.module';
-import { StFooterModule } from '@stratio/egeo';
 
 export const routes: Routes = [
    { path: '', redirectTo: 'demo/alerts-demo' },
@@ -29,7 +28,6 @@ export const routes: Routes = [
    imports: [
       CommonModule,
       SharedModule,
-      StFooterModule,
       RouterModule.forChild(routes)
    ],
    declarations: [DemoLayoutComponent]