+++ projects/egeo-demo/src/app/modules/home/home.ts
@@ -9,7 +9,6 @@
  * SPDX-License-Identifier: Apache-2.0.
  */
 import { Component } from '@angular/core';
-import { StFooterLink } from '@stratio/egeo';
 
 @Component({
    selector: 'demo-home',
@@ -18,33 +17,8 @@ import { StFooterLink } from '@stratio/egeo';
 })
 export class HomeComponent {
    public rightsText: string;
-   public links: StFooterLink[] = [];
 
    constructor() {
       this.rightsText = 'Copyright © 2018 Stratio';
-      this.links.push({
-            title: 'Github',
-            url: 'http://www.github.com/stratio',
-            icon: 'icon-github2'
-         },
-         {
-            title: 'Twitter',
-            url: 'https://twitter.com/stratiobd',
-            icon: 'icon-twitter'
-         },
-         {
-            title: 'Youtube',
-            url: 'https://www.youtube.com/channel/UCuzJA_GlT3TQaU5DyRwDOVw',
-            icon: 'icon-youtube'
-         },
-         {
-            title: 'Linkedin',
-            url: 'https://es.linkedin.com/company/stratiobd',
-            icon: 'icon-linkedin'
-         }, {
-            title: 'Slideshare',
-            url: 'https://es.slideshare.net/stratio',
-            icon: 'icon-slideshare'
-         });
    }
 }